* commit 'f8a64397efe93bd5a5e939fb81b1de4b359e2048':
  SIP-9526: Fixed check-style issue
  SIP-9526: minor ui tweeks
  SIP-9526: fixing build errors
  SIP-9526: enabling publish checkbox in admin module as its plugged to schedule
  SIP-9526: hiding publish buttons when system category is true
  SIP-9649 - BE: Do not allow to publish or delete analysis from system category folder
  SIP-9526 - Resolved Merge conflicts
  SIP-9526: Admin module changes for system category folders
  SIP-9526: FE changes for system category folders
  SIP-9578 - BE: Changes in sip-security/jwt/api to handle system category folder
  SIP-9578 - BE: Changes in sip-security/jwt/api to handle system category folder

+++ sip-services/sip-metadata-service/src/main/java/com/synchronoss/saw/analysis/controller/AnalysisController.java
@@ -1,7 +1,6 @@
 package com.synchronoss.saw.analysis.controller;
 
 import static com.synchronoss.saw.util.SipMetadataUtils.getTicket;
-import static com.synchronoss.saw.util.SipMetadataUtils.getToken;
 import static com.synchronoss.saw.util.SipMetadataUtils.validateTicket;
 import static com.synchronoss.sip.utils.SipCommonUtils.authValidation;
 import static com.synchronoss.sip.utils.SipCommonUtils.checkForPrivateCategory;
@@ -14,9 +13,9 @@ import com.synchronoss.saw.analysis.modal.Analysis;
 import com.synchronoss.saw.analysis.modal.AnalysisPrivileges;
 import com.synchronoss.saw.analysis.response.AnalysisResponse;
 import com.synchronoss.saw.analysis.service.AnalysisService;
-import com.synchronoss.saw.exceptions.SipAuthorizationException;
 import com.synchronoss.saw.util.SipMetadataUtils;
 import com.synchronoss.sip.utils.Privileges.PrivilegeNames;
+import com.synchronoss.sip.utils.SipCommonUtils;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
@@ -216,6 +215,14 @@ public class AnalysisController {
       analysisResponse.setMessage(HttpStatus.UNAUTHORIZED.getReasonPhrase());
       return analysisResponse;
     }
+
+    Long categoryId = analysis.getCategory() != null ? Long.valueOf(analysis.getCategory()) : 0L;
+    if (response != null && response.getStatus() != HttpStatus.UNAUTHORIZED.value()
+        && SipCommonUtils.haveSystemCategory(authTicket.getProducts(), categoryId)) {
+      analysisResponse.setMessage(HttpStatus.UNAUTHORIZED.getReasonPhrase());
+      return analysisResponse;
+    }
+
     analysisService.deleteAnalysis(id, authTicket);
     analysisResponse.setMessage("Analysis deleted successfully");
     analysisResponse.setAnalysisId(id);
